+++ ml-agents/mlagents/trainers/learn.py
@@ -38,7 +38,6 @@ from mlagents_envs.base_env import BaseEnv
 from mlagents.trainers.subprocess_env_manager import SubprocessEnvManager
 from mlagents_envs.side_channel.side_channel import SideChannel
 from mlagents_envs.side_channel.engine_configuration_channel import EngineConfig
-from mlagents_envs.exception import UnityEnvironmentException
 from mlagents_envs.timers import (
     hierarchical_timer,
     get_timer_tree,
@@ -357,7 +356,7 @@ def run_training(run_seed: int, options: RunOptions) -> None:
             os.path.join(base_path, options.run_id) if options.initialize_from else None
         )
         run_logs_dir = os.path.join(write_path, "run_logs")
-        port = options.base_port
+        port: Optional[int] = options.base_port
         # Check if directory exists
         handle_existing_directories(
             write_path, options.resume, options.force, maybe_init_path
@@ -382,7 +381,7 @@ def run_training(run_seed: int, options: RunOptions) -> None:
         StatsReporter.add_writer(console_writer)
 
         if options.env_path is None:
-            port = UnityEnvironment.DEFAULT_EDITOR_PORT
+            port = None
         env_factory = create_environment_factory(
             options.env_path,
             options.no_graphics,
@@ -505,17 +504,10 @@ def create_environment_factory(
     env_path: Optional[str],
     no_graphics: bool,
     seed: int,
-    start_port: int,
+    start_port: Optional[int],
     env_args: Optional[List[str]],
     log_folder: str,
 ) -> Callable[[int, List[SideChannel]], BaseEnv]:
-    if env_path is not None:
-        launch_string = UnityEnvironment.validate_environment_path(env_path)
-        if launch_string is None:
-            raise UnityEnvironmentException(
-                f"Couldn't launch the {env_path} environment. Provided filename does not match any environments."
-            )
-
     def create_unity_environment(
         worker_id: int, side_channels: List[SideChannel]
     ) -> UnityEnvironment: